chore - Format MailChimp the same as CRM integrations
